requirements.txt
jupyter
sentence-transformers
pgvector
django
python-decouple
llama-index
llama-index-vector-stores-postgres
sqlalchemy